Converted our database functions to take const char* arguments instead of char* arguments. Unimplemented our "Save two seconds before quitting" to our database. Because this is a new format (and it uses VBR by default) we have to try and discover exactly how we seek properly in this file.
Our "Jump to minutes" key currently is also not working for similar reasons as outlined above.
Fixed some include files on our UnixAudio file - yes, I have only tested this on FreeBSD as of yet.
Fixed a few warnings as well that had to do with not returning from the thread functions in the player.



git-svn-id: svn://mosedal.net/abf/code/trunk@422 3c19b6e9-aa5f-dd11-88a4-001302d38fef